# src/main/resources/openapi/semantic-api-v1.yaml
openapi: 3.0.3
info:
  title: Java Semantic Service Repository API
  version: v1
security:
  - ApiToken: []
paths:
...
  /v1/discovery/concepts:
    post:
      operationId: discoverConcepts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscoverConceptsRequest'
      responses:
        '200':
          $ref: '#/components/responses/ConceptDiscovery'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/AuthDisabled'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ConceptKindUnavailable'
        '500':
          $ref: '#/components/responses/MutationFailed'
  /v1/discovery/concepts/resolve:
    post:
      operationId: resolveConcept
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveConceptRequest'
      responses:
        '200':
          $ref: '#/components/responses/ConceptResolution'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/AuthDisabled'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/MutationFailed'
  /v1/discovery/type-members:
    post:
      operationId: discoverTypeMembers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscoverTypeMembersRequest'
      responses:
        '200':
          $ref: '#/components/responses/TypeMemberDiscovery'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/AuthDisabled'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/MutationFailed'
...
components:
  securitySchemes:
    ApiToken:
      type: apiKey
      in: header
      name: X-Api-Token
...
  responses:
...
    ConceptDiscovery:
      description: Revision-bound structured concept discovery candidates and recovery guidance
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DiscoverConceptsResponse'
    ConceptResolution:
      description: Revision-bound exact typed concept identity resolution
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResolveConceptResponse'
    ConceptKindUnavailable:
      description: Requested concept kind is declared but inactive in this composition
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ConceptKindUnavailableResponse'
    TypeMemberDiscovery:
      description: Revision-bound combined type member discovery page
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DiscoverTypeMembersResponse'
...
    BadRequest:
      description: Invalid repository id or request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Missing or wrong API token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    AuthDisabled:
      description: API token is not configured
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthDisabledErrorResponse'
    NotFound:
      description: Repository or analysis symbol is not configured
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: Repository is busy, immutable, not ready, or revision-mismatched
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
    MutationFailed:
      description: Sanitized service failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
  schemas:
...
    ResolveConceptRequest:
      type: object
      additionalProperties: false
      required:
        - repoId
        - expectedRevision
        - identity
      properties:
        repoId:
          type: string
          minLength: 1
          pattern: '^[a-z0-9][a-z0-9._-]{0,63}$'
        expectedRevision:
          type: string
          minLength: 1
          pattern: '^[0-9a-f]{40}$|^FIXTURE$'
        identity:
          $ref: '#/components/schemas/ConceptIdentityResponse'
...
    MethodTarget:
      type: object
      additionalProperties: false
      required:
        - sourceFile
        - packageName
        - className
        - methodName
        - parameterTypes
      properties:
        sourceFile:
          type: string
          minLength: 1
          maxLength: 1024
          pattern: '^[^\u0000-\u001F\u007F-\u009F]*[^\u0000-\u0020\u007F-\u009F\u1680\u2000-\u2006\u2008-\u200A\u2028-\u2029\u205F\u3000][^\u0000-\u001F\u007F-\u009F]*$'
          description: Repository-relative source path. The server additionally enforces normalized path segments.
        packageName:
          type: string
        className:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '^[\p{L}\p{Nl}\p{Sc}\p{Pc}][\p{L}\p{Nl}\p{Sc}\p{Pc}\p{Mn}\p{Mc}\p{Nd}]*(?:\.[\p{L}\p{Nl}\p{Sc}\p{Pc}][\p{L}\p{Nl}\p{Sc}\p{Pc}\p{Mn}\p{Mc}\p{Nd}]*)*$'
          description: Portable Unicode-category approximation of the server's dot-separated Java identifier validation.
        methodName:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '^[\p{L}\p{Nl}\p{Sc}\p{Pc}][\p{L}\p{Nl}\p{Sc}\p{Pc}\p{Mn}\p{Mc}\p{Nd}]*$'
          description: Portable Unicode-category approximation of the server's Java identifier validation.
        parameterTypes:
          type: array
          minItems: 0
          items:
            type: string
            minLength: 1
            pattern: '.*\S.*'
...
    DiscoverConceptsRequest:
      type: object
      additionalProperties: false
      required:
        - operator
        - repoId
        - expectedRevision
        - terms
        - kinds
      properties:
        operator:
          type: string
          enum:
            - ALL
        repoId:
          type: string
          minLength: 1
          pattern: '^[a-z0-9][a-z0-9._-]{0,63}$'
        expectedRevision:
          type: string
          minLength: 1
          pattern: '^[0-9a-f]{40}$|^FIXTURE$'
        terms:
          type: array
          minItems: 1
          maxItems: 4
          items:
            $ref: '#/components/schemas/ConceptSearchTerm'
        kinds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ConceptKind'
        packagePrefix:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '.*\S.*'
        offset:
          type: integer
          default: 0
          minimum: 0
        limit:
          type: integer
          default: 50
          minimum: 1
          maximum: 100
    ConceptSearchTerm:
      type: object
      additionalProperties: false
      required:
        - value
        - matchMode
      properties:
        value:
          type: string
          minLength: 2
          maxLength: 128
          pattern: '.*\S.*'
        matchMode:
          $ref: '#/components/schemas/ConceptMatchMode'
    ConceptKind:
      type: string
      enum:
        - TYPE
        - METHOD
        - FIELD
        - ANNOTATION_USAGE
        - TYPE_USAGE
        - API_ROUTE
        - MQ_DESTINATION
        - SCHEDULE
        - MAPPER_STATEMENT
        - SQL_IDENTIFIER
        - CONFIGURATION_KEY
        - OUTBOUND_API
        - MQ_PUBLISHER
        - ERROR_CONTRACT
        - ENUM_CONSTANT
    ConceptMatchMode:
      type: string
      enum:
        - TOKEN_EXACT
        - TOKEN_PREFIX
    DiscoverConceptsResponse:
      type: object
      additionalProperties: false
      required:
        - repoId
        - analyzedRevision
        - normalizedTerms
        - searchedKinds
        - supportedKinds
        - limitations
        - candidates
        - page
        - coverage
        - issueSummaries
        - availableFollowUps
        - unavailableFollowUps
      properties:
        repoId:
          type: string
        analyzedRevision:
          type: string
          pattern: '^[0-9a-f]{40}$|^FIXTURE$'
        normalizedTerms:
          type: array
          items:
            type: string
        searchedKinds:
          type: array
          items:
            $ref: '#/components/schemas/ConceptKind'
        supportedKinds:
          type: array
          items:
            $ref: '#/components/schemas/ConceptKind'
        limitations:
          type: array
          items:
            type: string
            enum:
              - SOURCE_BODY_NOT_SEARCHED
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/ConceptCandidateResponse'
        page:
          $ref: '#/components/schemas/ConceptPageResponse'
        coverage:
          $ref: '#/components/schemas/ConceptCoverageResponse'
        issueSummaries:
          type: array
          items:
            $ref: '#/components/schemas/ConceptIssueSummaryResponse'
        availableFollowUps:
          type: array
          items:
            $ref: '#/components/schemas/DiscoveryFollowUpResponse'
        unavailableFollowUps:
          type: array
          items:
            $ref: '#/components/schemas/UnavailableDiscoveryFollowUpResponse'
    ConceptCandidateResponse:
      type: object
      additionalProperties: false
      required:
        - identity
        - displayValue
        - matchedTerms
        - authority
        - evidence
        - availableFollowUps
      properties:
        identity:
          $ref: '#/components/schemas/ConceptIdentityResponse'
        displayValue:
          type: string
        matchedTerms:
          type: array
          items:
            type: string
        authority:
          type: string
          enum:
            - SYNTAX_RESOLVED
            - SYNTAX_DECLARED
            - FRAMEWORK_METADATA
            - WRITTEN_NAME_FALLBACK
            - METADATA_VALUE_UNRESOLVED
        details:
          $ref: '#/components/schemas/ConceptCandidateDetailsResponse'
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/ConceptEvidenceResponse'
        availableFollowUps:
          type: array
          items:
            $ref: '#/components/schemas/ConceptCandidateFollowUpResponse'
...
    ConceptEvidenceResponse:
      type: object
      additionalProperties: false
      required:
        - identity
      properties:
        identity:
          $ref: '#/components/schemas/ConceptIdentityResponse'
    ConceptCandidateDetailsResponse:
      oneOf:
        - $ref: '#/components/schemas/FieldDetailsResponse'
        - $ref: '#/components/schemas/MapperStatementDetailsResponse'
      discriminator:
        propertyName: kind
        mapping:
          FIELD: '#/components/schemas/FieldDetailsResponse'
          MAPPER_STATEMENT: '#/components/schemas/MapperStatementDetailsResponse'
...
    ConceptIdentityResponse:
      oneOf:
        - $ref: '#/components/schemas/TypeConceptIdentityResponse'
        - $ref: '#/components/schemas/MethodConceptIdentityResponse'
        - $ref: '#/components/schemas/FieldConceptIdentityResponse'
        - $ref: '#/components/schemas/AnnotationUsageConceptIdentityResponse'
        - $ref: '#/components/schemas/TypeUsageConceptIdentityResponse'
        - $ref: '#/components/schemas/ApiRouteConceptIdentityResponse'
        - $ref: '#/components/schemas/MqDestinationConceptIdentityResponse'
        - $ref: '#/components/schemas/ScheduleConceptIdentityResponse'
        - $ref: '#/components/schemas/MapperStatementConceptIdentityResponse'
        - $ref: '#/components/schemas/MapperStatementVariantConceptIdentityResponse'
      discriminator:
        propertyName: kind
        mapping:
          TYPE: '#/components/schemas/TypeConceptIdentityResponse'
          METHOD: '#/components/schemas/MethodConceptIdentityResponse'
          FIELD: '#/components/schemas/FieldConceptIdentityResponse'
          ANNOTATION_USAGE: '#/components/schemas/AnnotationUsageConceptIdentityResponse'
          TYPE_USAGE: '#/components/schemas/TypeUsageConceptIdentityResponse'
          API_ROUTE: '#/components/schemas/ApiRouteConceptIdentityResponse'
          MQ_DESTINATION: '#/components/schemas/MqDestinationConceptIdentityResponse'
          SCHEDULE: '#/components/schemas/ScheduleConceptIdentityResponse'
          MAPPER_STATEMENT: '#/components/schemas/MapperStatementConceptIdentityResponse'
          MAPPER_STATEMENT_VARIANT: '#/components/schemas/MapperStatementVariantConceptIdentityResponse'
    TypeConceptIdentityResponse:
      type: object
      additionalProperties: false
      required: [kind, sourceFile, packageName, className]
      properties:
        kind: {type: string, enum: [TYPE]}
        sourceFile: {type: string}
        packageName: {type: string}
        className: {type: string}
    MethodConceptIdentityResponse:
      type: object
      additionalProperties: false
      required: [kind, target]
      properties:
        kind: {type: string, enum: [METHOD]}
        target: {$ref: '#/components/schemas/MethodTarget'}
    FieldConceptIdentityResponse:
      type: object
      additionalProperties: false
      required: [kind, sourceFile, ownerPackageName, ownerClassName, fieldName]
      properties:
        kind: {type: string, enum: [FIELD]}
        sourceFile: {type: string}
        ownerPackageName: {type: string}
        ownerClassName: {type: string}
        fieldName: {type: string}
    ResolveConceptResponse:
      type: object
      additionalProperties: false
      required: [repoId, analyzedRevision, candidate]
      properties:
        repoId:
          type: string
        analyzedRevision:
          type: string
          pattern: '^[0-9a-f]{40}$|^FIXTURE$'
        candidate:
          $ref: '#/components/schemas/ConceptCandidateResponse'
    AnnotationUsageConceptIdentityResponse:
      type: object
      additionalProperties: false
      required: [kind, declaration, annotationType]
      properties:
        kind: {type: string, enum: [ANNOTATION_USAGE]}
        declaration: {$ref: '#/components/schemas/DeclarationSubjectResponse'}
        annotationType: {$ref: '#/components/schemas/AnnotationTypeResponse'}
    TypeUsageConceptIdentityResponse:
      type: object
      additionalProperties: false
      required: [kind, owner, location, path, referencedType]
      properties:
        kind: {type: string, enum: [TYPE_USAGE]}
        owner: {$ref: '#/components/schemas/DeclarationSubjectResponse'}
        location: {$ref: '#/components/schemas/TypeUsageLocationResponse'}
        path:
          type: array
          items: {$ref: '#/components/schemas/TypeUsagePathResponse'}
        referencedType: {$ref: '#/components/schemas/ReferencedTypeResponse'}
    ApiRouteConceptIdentityResponse:
      type: object
      additionalProperties: false
      required: [kind, target, httpVerb, route]
      properties:
        kind: {type: string, enum: [API_ROUTE]}
        target: {$ref: '#/components/schemas/MethodTarget'}
        httpVerb: {type: string}
        route: {type: string}
    MqDestinationConceptIdentityResponse:
      type: object
      additionalProperties: false
      required: [kind, target, broker, destination]
      properties:
        kind: {type: string, enum: [MQ_DESTINATION]}
        target: {$ref: '#/components/schemas/MethodTarget'}
        broker: {type: string, enum: [RABBIT, KAFKA]}
        destination: {type: string}
    ScheduleConceptIdentityResponse:
      type: object
      additionalProperties: false
      required: [kind, target, triggerKind]
      properties:
        kind: {type: string, enum: [SCHEDULE]}
        target: {$ref: '#/components/schemas/MethodTarget'}
        triggerKind: {type: string, enum: [CRON, FIXED_DELAY, FIXED_RATE, JOB_HANDLER, UNSPECIFIED]}
        triggerValue: {type: string}
    MapperStatementConceptIdentityResponse:
      type: object
      additionalProperties: false
      required: [kind, statement]
      properties:
        kind: {type: string, enum: [MAPPER_STATEMENT]}
        statement: {$ref: '#/components/schemas/MapperStatementKeyResponse'}
    MapperStatementVariantConceptIdentityResponse:
      type: object
      additionalProperties: false
      required: [kind, variant]
      properties:
        kind: {type: string, enum: [MAPPER_STATEMENT_VARIANT]}
        variant: {$ref: '#/components/schemas/MapperStatementVariantResponse'}
...
    DeclarationSubjectResponse:
      oneOf:
        - $ref: '#/components/schemas/TypeDeclarationSubjectResponse'
        - $ref: '#/components/schemas/ResolvedMethodDeclarationSubjectResponse'
        - $ref: '#/components/schemas/UnresolvedMethodDeclarationSubjectResponse'
        - $ref: '#/components/schemas/FieldDeclarationSubjectResponse'
      discriminator:
        propertyName: kind
        mapping:
          TYPE: '#/components/schemas/TypeDeclarationSubjectResponse'
          METHOD: '#/components/schemas/ResolvedMethodDeclarationSubjectResponse'
          METHOD_UNRESOLVED: '#/components/schemas/UnresolvedMethodDeclarationSubjectResponse'
          FIELD: '#/components/schemas/FieldDeclarationSubjectResponse'
...
    AnnotationTypeResponse:
      oneOf:
        - $ref: '#/components/schemas/ResolvedAnnotationTypeResponse'
        - $ref: '#/components/schemas/UnresolvedAnnotationTypeResponse'
      discriminator:
        propertyName: status
        mapping:
          RESOLVED: '#/components/schemas/ResolvedAnnotationTypeResponse'
          UNRESOLVED: '#/components/schemas/UnresolvedAnnotationTypeResponse'
...
    TypeUsageLocationResponse:
      type: object
      additionalProperties: false
      required: [slot, index]
      properties:
        slot: {type: string, enum: [IMPLEMENTED_TYPE, EXTENDED_TYPE, METHOD_PARAMETER, METHOD_RETURN, METHOD_BODY_OR_ANNOTATION_MEMBER, FIELD_DECLARATION]}
        index: {type: integer, minimum: 0}
    ReferencedTypeResponse:
      type: object
      additionalProperties: false
      required: [javaType, arrayDimensions]
      properties:
        javaType: {$ref: '#/components/schemas/JavaTypeIdentityResponse'}
        arrayDimensions: {type: integer, minimum: 0}
    TypeUsagePathResponse:
      oneOf:
        - $ref: '#/components/schemas/TypeArgumentPathResponse'
        - $ref: '#/components/schemas/WildcardExtendsBoundPathResponse'
        - $ref: '#/components/schemas/WildcardSuperBoundPathResponse'
        - $ref: '#/components/schemas/TypeVariableBoundPathResponse'
      discriminator:
        propertyName: kind
        mapping:
          TYPE_ARGUMENT: '#/components/schemas/TypeArgumentPathResponse'
          WILDCARD_EXTENDS_BOUND: '#/components/schemas/WildcardExtendsBoundPathResponse'
          WILDCARD_SUPER_BOUND: '#/components/schemas/WildcardSuperBoundPathResponse'
          TYPE_VARIABLE_BOUND: '#/components/schemas/TypeVariableBoundPathResponse'
...
    ConceptPageResponse:
      type: object
      additionalProperties: false
      required:
        - offset
        - limit
        - returnedCount
        - totalCount
        - hasMore
      properties:
        offset:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
          maximum: 100
        returnedCount:
          type: integer
          minimum: 0
        totalCount:
          type: integer
          minimum: 0
        hasMore:
          type: boolean
    ConceptCoverageResponse:
      type: object
      additionalProperties: false
      required:
        - status
        - scannedFileCount
        - extractedFileCount
        - syntaxFailedFileCount
      properties:
        status:
          type: string
          enum:
            - COMPLETE
            - PARTIAL
        scannedFileCount:
          type: integer
          minimum: 0
        extractedFileCount:
          type: integer
          minimum: 0
        syntaxFailedFileCount:
          type: integer
          minimum: 0
    ConceptIssueSummaryResponse:
      type: object
      additionalProperties: false
      required:
        - reason
        - count
      properties:
        reason:
          type: string
          enum:
            - MQ_DESTINATION_UNRESOLVED
            - SCHEDULE_TRIGGER_VALUE_UNRESOLVED
        count:
          type: integer
          minimum: 1
    UnavailableDiscoveryFollowUpResponse:
      type: object
      additionalProperties: false
      required:
        - reason
        - recommendedAction
      properties:
        reason:
          type: string
          enum:
            - NO_MATCHING_STRUCTURED_CONCEPT
            - SEARCH_INCOMPLETE
        recommendedAction:
          type: string
          enum:
            - REFINE_TERMS_KINDS_OR_PACKAGE_FILTERS
            - FIX_SOURCE_OR_RETRY
    DiscoverTypeMembersRequest:
      type: object
      additionalProperties: false
      required:
        - repoId
        - expectedRevision
        - sourceFile
        - fullyQualifiedName
        - memberKinds
      properties:
        repoId:
          type: string
          minLength: 1
          pattern: '^[a-z0-9][a-z0-9._-]{0,63}$'
        expectedRevision:
          type: string
          minLength: 1
          pattern: '^[0-9a-f]{40}$|^FIXTURE$'
        sourceFile:
          type: string
          minLength: 1
          maxLength: 1024
          pattern: '^[^\u0000-\u001F\u007F-\u009F]*[^\u0000-\u0020\u007F-\u009F\u1680\u2000-\u2006\u2008-\u200A\u2028-\u2029\u205F\u3000][^\u0000-\u001F\u007F-\u009F]*$'
          description: Repository-relative source path. The server additionally enforces normalized path segments.
        fullyQualifiedName:
          type: string
          minLength: 1
          maxLength: 1024
          pattern: '.*\S.*'
        memberKinds:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/TypeMemberKind'
        namePrefix:
          type: string
          nullable: true
          maxLength: 255
        offset:
          type: integer
          default: 0
          minimum: 0
        limit:
          type: integer
          default: 50
          minimum: 1
          maximum: 100
    TypeMemberKind:
      type: string
      enum:
        - METHOD
        - FIELD
    DiscoverTypeMembersResponse:
      type: object
      additionalProperties: false
      required:
        - repoId
        - analyzedRevision
        - sourceFile
        - fullyQualifiedName
        - typeKind
        - annotations
        - implementedTypes
        - extendedTypes
        - members
        - page
        - coverage
        - availableFollowUps
      properties:
        repoId:
          type: string
        analyzedRevision:
          type: string
          pattern: '^[0-9a-f]{40}$|^FIXTURE$'
        sourceFile:
          type: string
        fullyQualifiedName:
          type: string
        typeKind:
          type: string
        annotations:
          type: array
          items:
            type: string
        implementedTypes:
          type: array
          items:
            type: string
        extendedTypes:
          type: array
          items:
            type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/TypeMemberResponse'
        page:
          $ref: '#/components/schemas/ConceptPageResponse'
        coverage:
          $ref: '#/components/schemas/ConceptCoverageResponse'
        availableFollowUps:
          type: array
          items:
            $ref: '#/components/schemas/DiscoveryFollowUpResponse'
    TypeMemberResponse:
      oneOf:
        - $ref: '#/components/schemas/MethodTypeMemberResponse'
        - $ref: '#/components/schemas/FieldTypeMemberResponse'
      discriminator:
        propertyName: kind
        mapping:
          METHOD: '#/components/schemas/MethodTypeMemberResponse'
          FIELD: '#/components/schemas/FieldTypeMemberResponse'
...
    ConceptCandidateFollowUpResponse:
      type: object
      additionalProperties: false
      required:
        - operation
        - api
        - request
      properties:
        operation:
          type: string
          enum:
            - GET_METHOD_SOURCE
            - GET_METHOD_SQL
            - ANALYZE_OUTGOING_CALL_GRAPH
            - ANALYZE_INCOMING_CALL_GRAPH
            - DISCOVER_METHOD_IMPLEMENTATIONS
            - RESOLVE_CONCEPT
            - GET_TYPE_MEMBERS
            - GET_NEXT_PAGE
        api:
          $ref: '#/components/schemas/DiscoveryFollowUpApiResponse'
        request:
          $ref: '#/components/schemas/ConceptCandidateFollowUpRequestResponse'
    DiscoveryFollowUpResponse:
      type: object
      additionalProperties: false
      required:
        - operation
        - api
        - request
      properties:
        operation:
          type: string
          enum:
            - GET_METHOD_SOURCE
            - GET_METHOD_SQL
            - GET_MAPPER_FRAGMENT
            - GET_METHOD_SOURCE_SEGMENT
            - GET_METHOD_SQL_SEGMENT
            - GET_MAPPER_FRAGMENT_SEGMENT
            - ANALYZE_OUTGOING_CALL_GRAPH
            - ANALYZE_INCOMING_CALL_GRAPH
            - DISCOVER_METHOD_IMPLEMENTATIONS
            - RESOLVE_CONCEPT
            - GET_TYPE_MEMBERS
            - GET_NEXT_PAGE
            - RESOLVE_SOURCE_SYMBOL
        api:
          $ref: '#/components/schemas/DiscoveryFollowUpApiResponse'
        request:
          $ref: '#/components/schemas/DiscoveryFollowUpRequestResponse'
    DiscoveryFollowUpApiResponse:
      type: object
      additionalProperties: false
      required:
        - method
        - path
        - operationId
      properties:
        method:
          type: string
          enum:
            - POST
        path:
          type: string
        operationId:
          type: string
...
    DiscoveryFollowUpRequestResponse:
      oneOf:
        - $ref: '#/components/schemas/MethodTargetFollowUpRequestResponse'
        - $ref: '#/components/schemas/GetMapperFragmentRequestResponse'
        - $ref: '#/components/schemas/GetMethodSourceSegmentRequestResponse'
        - $ref: '#/components/schemas/GetMapperStatementSegmentRequestResponse'
        - $ref: '#/components/schemas/GetMapperFragmentSegmentRequestResponse'
        - $ref: '#/components/schemas/AnalyzeCallGraphRequestResponse'
        - $ref: '#/components/schemas/DiscoverMethodImplementationsRequestResponse'
        - $ref: '#/components/schemas/ResolveConceptRequestResponse'
        - $ref: '#/components/schemas/ConceptSearchPageRequestResponse'
        - $ref: '#/components/schemas/GetTypeMembersRequestResponse'
        - $ref: '#/components/schemas/DiscoverTypeMembersRequestResponse'
        - $ref: '#/components/schemas/ResolveSourceSymbolRequestResponse'
...
    ConceptKindUnavailableResponse:
      type: object
      additionalProperties: false
      required:
        - errorCode
        - message
        - unavailableKinds
        - supportedKinds
        - requestId
      properties:
        errorCode:
          type: string
          enum:
            - CONCEPT_KIND_UNAVAILABLE
        message:
          type: string
        unavailableKinds:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ConceptKind'
        supportedKinds:
          type: array
          items:
            $ref: '#/components/schemas/ConceptKind'
        requestId:
          type: string
    ApiErrorResponse:
      type: object
      additionalProperties: false
      required:
        - errorCode
        - message
        - candidates
      properties:
        errorCode:
          type: string
          enum:
            - REQUEST_INVALID
            - SEMANTIC_UNAUTHORIZED
            - REPOSITORY_NOT_FOUND
            - REPOSITORY_NOT_READY
            - REPOSITORY_REVISION_MISMATCH
            - SEMANTIC_BINDING_AMBIGUOUS
            - SEMANTIC_TARGET_NOT_FOUND
            - EXACT_CONTENT_NOT_FOUND
            - TYPE_MEMBER_TYPE_NOT_FOUND
            - SEMANTIC_BINDING_UNRESOLVED
            - IMPLEMENTATION_TARGET_UNSUPPORTED
            - SEMANTIC_PROTOCOL_ERROR
            - SEMANTIC_ENGINE_START_FAILED
            - SEMANTIC_REQUEST_TIMEOUT
            - INTERNAL_ERROR
        message:
          type: string
        repoId:
          type: string
          nullable: true
        currentRevision:
          type: string
          nullable: true
          pattern: '^[0-9a-f]{40}$|^FIXTURE$'
        expectedRevision:
          type: string
          nullable: true
          pattern: '^[0-9a-f]{40}$|^FIXTURE$'
        target:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/MethodTarget'
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/MethodTarget'
        requestId:
          type: string
          nullable: true
    AuthDisabledErrorResponse:
      type: object
      additionalProperties: false
      required:
        - errorCode
        - message
        - candidates
      properties:
        errorCode:
          type: string
          enum:
            - SEMANTIC_AUTH_DISABLED
        message:
          type: string
        repoId:
          type: string
          nullable: true
        expectedRevision:
          type: string
          nullable: true
          pattern: '^[0-9a-f]{40}$|^FIXTURE$'
        currentRevision:
          type: string
          nullable: true
          pattern: '^[0-9a-f]{40}$|^FIXTURE$'
        target:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/MethodTarget'
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/MethodTarget'
        requestId:
          type: string
          nullable: true